requirements.txt
streamlit
pandas
plotly
kaleido
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib